Update: Merged callback and callback_typed. callback now behaves as callback_typed did. Update: Rewrote tests to match above change.
