+++ dash/__init__.py
@@ -19,7 +19,7 @@ from . import html  # noqa: F401,E402
 from . import dash_table  # noqa: F401,E402
 from .version import __version__  # noqa: F401,E402
 from ._callback_context import callback_context, set_props  # noqa: F401,E402
-from ._callback import callback, clientside_callback, typed_callback  # noqa: F401,E402
+from ._callback import callback, clientside_callback  # noqa: F401,E402
 from ._get_app import get_app  # noqa: F401,E402
 from ._get_paths import (  # noqa: F401,E402
     get_asset_url,
@@ -77,7 +77,6 @@ __all__ = [
     "callback_context",
     "set_props",
     "callback",
-    "typed_callback",
     "get_app",
     "get_asset_url",
     "get_relative_path",
