+++ tests/integration/callbacks/test_callback_decorator.py
@@ -1,44 +1,44 @@
-"""Integration tests for typed_callback functionality."""
+"""Integration tests for callback decorator behavior."""
 import dash
-from dash import Input, Output, State, dcc, html, typed_callback
+from dash import Input, Output, State, callback, dcc, html
 
 
-def test_tcb001_basic_typed_callback(dash_duo):
-    """Test that typed_callback works identically to callback for basic usage."""
+def test_cb001_callback_and_dash_callback_equivalence(dash_duo):
+    """Test that callback and dash.callback behave equivalently for basic usage."""
     app = dash.Dash(__name__)
 
     app.layout = html.Div(
         [
             dcc.Input(id="input", value="initial"),
-            html.Div(id="output-typed"),
-            html.Div(id="output-regular"),
+            html.Div(id="output-callback"),
+            html.Div(id="output-dash-callback"),
         ]
     )
 
-    @typed_callback(Output("output-typed", "children"), Input("input", "value"))
-    def update_typed(value):
-        return f"Typed: {value}"
+    @callback(Output("output-callback", "children"), Input("input", "value"))
+    def update_callback(value):
+        return f"Callback: {value}"
 
-    @dash.callback(Output("output-regular", "children"), Input("input", "value"))
-    def update_regular(value):
-        return f"Regular: {value}"
+    @dash.callback(Output("output-dash-callback", "children"), Input("input", "value"))
+    def update_dash_callback(value):
+        return f"Dash callback: {value}"
 
     dash_duo.start_server(app)
-    dash_duo.wait_for_text_to_equal("#output-typed", "Typed: initial")
-    dash_duo.wait_for_text_to_equal("#output-regular", "Regular: initial")
+    dash_duo.wait_for_text_to_equal("#output-callback", "Callback: initial")
+    dash_duo.wait_for_text_to_equal("#output-dash-callback", "Dash callback: initial")
 
     input_element = dash_duo.find_element("#input")
     dash_duo.clear_input(input_element)
     input_element.send_keys("test")
 
-    dash_duo.wait_for_text_to_equal("#output-typed", "Typed: test")
-    dash_duo.wait_for_text_to_equal("#output-regular", "Regular: test")
+    dash_duo.wait_for_text_to_equal("#output-callback", "Callback: test")
+    dash_duo.wait_for_text_to_equal("#output-dash-callback", "Dash callback: test")
 
     assert not dash_duo.get_logs()
 
 
-def test_tcb002_typed_callback_multi_input(dash_duo):
-    """Test typed_callback with multiple inputs."""
+def test_cb002_callback_multi_input(dash_duo):
+    """Test callback with multiple inputs."""
     app = dash.Dash(__name__)
 
     app.layout = html.Div(
@@ -49,7 +49,7 @@ def test_tcb002_typed_callback_multi_input(dash_duo):
         ]
     )
 
-    @typed_callback(
+    @callback(
         Output("output", "children"),
         Input("input1", "value"),
         Input("input2", "value"),
@@ -69,8 +69,8 @@ def test_tcb002_typed_callback_multi_input(dash_duo):
     assert not dash_duo.get_logs()
 
 
-def test_tcb003_typed_callback_with_state(dash_duo):
-    """Test typed_callback with State dependencies."""
+def test_cb003_callback_with_state(dash_duo):
+    """Test callback with State dependencies."""
     app = dash.Dash(__name__)
 
     app.layout = html.Div(
@@ -82,7 +82,7 @@ def test_tcb003_typed_callback_with_state(dash_duo):
         ]
     )
 
-    @typed_callback(
+    @callback(
         Output("output", "children"),
         Input("button", "n_clicks"),
         State("input1", "value"),
@@ -105,8 +105,8 @@ def test_tcb003_typed_callback_with_state(dash_duo):
     assert not dash_duo.get_logs()
 
 
-def test_tcb004_typed_callback_multi_output(dash_duo):
-    """Test typed_callback with multiple outputs."""
+def test_cb004_callback_multi_output(dash_duo):
+    """Test callback with multiple outputs."""
     app = dash.Dash(__name__)
 
     app.layout = html.Div(
@@ -118,7 +118,7 @@ def test_tcb004_typed_callback_multi_output(dash_duo):
         ]
     )
 
-    @typed_callback(
+    @callback(
         Output("output1", "children"),
         Output("output2", "children"),
         Output("output3", "children"),
@@ -143,8 +143,8 @@ def test_tcb004_typed_callback_multi_output(dash_duo):
     assert not dash_duo.get_logs()
 
 
-def test_tcb005_typed_callback_prevent_initial_call(dash_duo):
-    """Test typed_callback with prevent_initial_call parameter."""
+def test_cb005_callback_prevent_initial_call(dash_duo):
+    """Test callback with prevent_initial_call parameter."""
     app = dash.Dash(__name__)
 
     app.layout = html.Div(
@@ -154,7 +154,7 @@ def test_tcb005_typed_callback_prevent_initial_call(dash_duo):
         ]
     )
 
-    @typed_callback(
+    @callback(
         Output("output", "children"),
         Input("input", "value"),
         prevent_initial_call=True,
@@ -163,7 +163,6 @@ def test_tcb005_typed_callback_prevent_initial_call(dash_duo):
         return f"Updated: {value}"
 
     dash_duo.start_server(app)
-    # Should remain "default" because prevent_initial_call=True
     dash_duo.wait_for_text_to_equal("#output", "default")
 
     input_element = dash_duo.find_element("#input")
@@ -174,8 +173,8 @@ def test_tcb005_typed_callback_prevent_initial_call(dash_duo):
     assert not dash_duo.get_logs()
 
 
-def test_tcb006_typed_callback_mixed_with_regular(dash_duo):
-    """Test that typed_callback and regular callback can coexist."""
+def test_cb006_callback_global_and_app_callback_mix(dash_duo):
+    """Test that callback, dash.callback, and app.callback can coexist."""
     app = dash.Dash(__name__)
 
     app.layout = html.Div(
@@ -188,41 +187,37 @@ def test_tcb006_typed_callback_mixed_with_regular(dash_duo):
         ]
     )
 
-    # Using typed_callback (global style)
-    @typed_callback(Output("output1", "children"), Input("input", "value"))
+    @callback(Output("output1", "children"), Input("input", "value"))
     def update_1(value):
-        return f"Typed global: {value}"
+        return f"Callback global: {value}"
 
-    # Using dash.callback (global style)
     @dash.callback(Output("output2", "children"), Input("input", "value"))
     def update_2(value):
-        return f"Regular global: {value}"
+        return f"Dash callback global: {value}"
 
-    # Using app.callback (app instance style)
     @app.callback(Output("output3", "children"), Input("input", "value"))
     def update_3(value):
         return f"App callback: {value}"
 
-    # Another typed_callback
-    @typed_callback(Output("output4", "children"), Input("input", "value"))
+    @callback(Output("output4", "children"), Input("input", "value"))
     def update_4(value):
-        return f"Typed global 2: {value}"
+        return f"Callback global 2: {value}"
 
     dash_duo.start_server(app)
 
     input_element = dash_duo.find_element("#input")
     input_element.send_keys("test")
 
-    dash_duo.wait_for_text_to_equal("#output1", "Typed global: test")
-    dash_duo.wait_for_text_to_equal("#output2", "Regular global: test")
+    dash_duo.wait_for_text_to_equal("#output1", "Callback global: test")
+    dash_duo.wait_for_text_to_equal("#output2", "Dash callback global: test")
     dash_duo.wait_for_text_to_equal("#output3", "App callback: test")
-    dash_duo.wait_for_text_to_equal("#output4", "Typed global 2: test")
+    dash_duo.wait_for_text_to_equal("#output4", "Callback global 2: test")
 
     assert not dash_duo.get_logs()
 
 
-def test_tcb007_typed_callback_with_no_update(dash_duo):
-    """Test typed_callback with no_update."""
+def test_cb007_callback_with_no_update(dash_duo):
+    """Test callback with no_update."""
     app = dash.Dash(__name__)
 
     app.layout = html.Div(
@@ -233,7 +228,7 @@ def test_tcb007_typed_callback_with_no_update(dash_duo):
         ]
     )
 
-    @typed_callback(
+    @callback(
         Output("output1", "children"),
         Output("output2", "children"),
         Input("input", "value"),
@@ -242,8 +237,7 @@ def test_tcb007_typed_callback_with_no_update(dash_duo):
         num = int(value) if value and value.isdigit() else 0
         if num % 2 == 0:
             return f"Even: {num}", dash.no_update
-        else:
-            return dash.no_update, f"Odd: {num}"
+        return dash.no_update, f"Odd: {num}"
 
     dash_duo.start_server(app)
     dash_duo.wait_for_text_to_equal("#output1", "Even: 0")
